requirements.txt
Pillow
PyYAML
opencv-python
numpy
deepface
